requirements_dev.txt
# Generic python dependencies
black==24.10.0
mypy==1.13.0
pip==24.3.1
pre-commit==4.0.1
ruff==0.8.1

# Home Assistant testing dependencies
pytest-homeassistant-custom-component==0.13.186
syrupy>=4.6.1

# Component dependencies
hassil>=2.0.4
home-assistant-intents>=2024.4.3
